projects/capstone/starter/requirements.txt
astroid==2.2.5
autopep8==1.5.2
cffi==1.14.0
Click==7.0
colorama==0.4.3
cryptography==2.9.2
ecdsa==0.13.2
Flask==1.0.2
Flask-Cors==3.0.8
Flask-SQLAlchemy==2.4.0
future==0.17.1
isort==4.3.18
itsdangerous==1.1.0
Jinja2==2.10.1
jwt==1.0.0
lazy-object-proxy==1.4.0
MarkupSafe==1.1.1
mccabe==0.6.1
pyasn1==0.4.8
pycodestyle==2.6.0
pycparser==2.20
pylint==2.5.2
python-dotenv==0.13.0
python-jose==3.1.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.3
toml==0.10.1
typed-ast==1.3.5
Werkzeug==0.15.2
wrapt==1.11.1
